dispatcher/requirements.txt
fastapi
uvicorn
requests
pytest
httpx
pymongo
python-dotenv